# influxdata: include line and column in decoder errors